ref it's work

+++ src/IRecipe.cpp
@@ -28,7 +28,7 @@ void RecipeOut::JsonLoad(DataBase& db, const rapidjson::Value& val)
   if (val.Capacity() >= 2)
     count = val[1].GetInt();
   if (val.Capacity() >= 3)
-    chance = val[2].GetDouble();
+    chance = static_cast<float>(val[2].GetDouble());
 }
 
 bool RecipeOut::operator==(const RecipeOut& other) const
